minimize changes for sshd PermitEmptyPasswords
From the STIG V1R4 RHEL-07-010300:

If no line, a commented line, or a line indicating the value "no" is
returned, the required value is set.
If the required value is not set, this is a finding.

Also, from the default sshd_config file:

The strategy used for options in the default sshd_config shipped with
OpenSSH is to specify options with their default value where
possible, but leave them commented.  Uncommented options override the
default value.